refactor: move test collectors to root directory
- Moved collector files from subdirectories to root:
  - core/test_core.py -> test_core.py
  - handlers/test_handlers.py -> test_handlers.py
  - media/test_media.py -> test_media.py
  - sync/test_sync.py -> test_sync.py
- Updated test_all.py to use new import paths
- Follows test structure specified in testing.md
