+++ scripts/sync_engine/tests/test_all.py
@@ -1,14 +1,14 @@
 """Main test collector - imports all test categories"""
 
 # Core tests
-from sync_engine.tests.core.test_core import (
+from sync_engine.tests.test_core import (
     TestAtomicOperations,
     TestChangeDetection,
     TestConfigHandling
 )
 
 # Media tests
-from sync_engine.tests.media.test_media import (
+from sync_engine.tests.test_media import (
     TestMediaProcessing,
     TestMediaSync,
     TestMediaReferences,
@@ -17,7 +17,7 @@ from sync_engine.tests.media.test_media import (
 )
 
 # Sync tests
-from sync_engine.tests.sync.test_sync import (
+from sync_engine.tests.test_sync import (
     TestBasicSync,
     TestSyncErrors,
     TestSyncCleanup,
@@ -27,7 +27,7 @@ from sync_engine.tests.sync.test_sync import (
 )
 
 # Handler tests
-from sync_engine.tests.handlers.test_handlers import TestPostHandling
+from sync_engine.tests.test_handlers import TestPostHandling
 
 # This file serves as the main collector of all tests
 # Running this file will run the complete test suite
